+++ app/Filament/Resources/SettlementResource/Pages/ListSettlements.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace App\Filament\Resources\SettlementResource\Pages;
+
+use App\Filament\Resources\SettlementResource;
+use Filament\Resources\Pages\ListRecords;
+
+class ListSettlements extends ListRecords
+{
+    protected static string $resource = SettlementResource::class;
+
+    protected function getHeaderActions(): array
+    {
+        return [];
+    }
+}